requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
anthropic
google-generativeai
mistralai
python-multipart
tiktoken